bind colors to processors and don't let the chart choose them automatically
